requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.32.2
python-dotenv==1.0.1
httpx==0.27.0

# NLP / processing
scikit-learn==1.4.2
numpy==1.26.4
pytrends>=4.9.2

# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.2.0

#Phase 3 — LinkedIn via Apify (install when ready)
apify-client>=1.7.0

#Phase 4 — Caching & Persistence (install when ready)
redis>=5.0.0
psycopg2-binary>=2.9.9

#Phase 5 — Output & Notifications (install when ready)
notion-client>=2.2.1
sendgrid>=6.11.0